cast result of shmat/shmget to a long rather than an int as a long is generally 	the same size as a void * but an int may be smaller, using an int 	causes a compile-time error on amd64

+++ src/sharedmemory.C
@@ -93,7 +93,7 @@ bool sharedmemory::create(key_t key, size_t size, mode_t permissions) {
 		// attach to the segment, remove the
 		// segment and return 0 on failure
 		shmptr=shmat(shmid,0,0);
-		if ((int)shmptr==-1) {
+		if ((long)shmptr==-1) {
 			forceRemove();
 			return false;
 		}
@@ -125,7 +125,7 @@ bool sharedmemory::attach(key_t key) {
 	// to see if it's not -1, but allow that it could very well be less
 	// than -1 and still be valid.
 	return ((shmid=shmget(key,0,0))!=-1 &&
-			(int)(shmptr=shmat(shmid,0,0))!=-1);
+			(long)(shmptr=shmat(shmid,0,0))!=-1);
 }
 
 bool sharedmemory::createOrAttach(key_t key, size_t size, mode_t permissions) {
@@ -139,7 +139,7 @@ bool sharedmemory::createOrAttach(key_t key, size_t size, mode_t permissions) {
 		// attach to the segment, remove the
 		// segment and return 0 on failure
 		shmptr=shmat(shmid,0,0);
-		if ((int)shmptr==-1) {
+		if ((long)shmptr==-1) {
 			forceRemove();
 			return false;
 		}
@@ -152,7 +152,7 @@ bool sharedmemory::createOrAttach(key_t key, size_t size, mode_t permissions) {
 
 		// attach to the segment, return 1 on success and 0 on failure
 		shmptr=shmat(shmid,0,0);
-		return ((int)shmptr!=-1);
+		return ((long)shmptr!=-1);
 
 	}
